requirements.txt
pytest
appium-python-client
pytest-allure-adaptor
allure-pytest
